Refactor index.html.
